log_management_system/requirements.txt
django
djangorestframework
django-cors-headers
elasticsearch
gunicorn
drf-spectacular